Add files via upload
Updated the code to add in a method to get the funding data from the matching ORCiD record. The funding data does not pull through the funding amount though so this may need extra work. 

Also working on pulling through the name of the user from their ORCiD record to save them having to type it in.